[BACKLOG-194] - Document Public Rest Endpoints for Repository File Structure / Management. Refactor for rest endpoint PUT /repo/files/path/rename(FileResource.doRename)
